Remove kotlin-test-junit library
+ Update libraries

+++ data/src/test/kotlin/com/lightteam/modpeide/data/ExampleUnitTest.kt
@@ -21,8 +21,9 @@ import org.junit.Assert.assertEquals
 import org.junit.Test
 
 class ExampleUnitTest {
+
     @Test
     fun addition_isCorrect() {
-        assertEquals(4, (2 + 2).toLong())
+        assertEquals(4, 2 + 2)
     }
 }
